requirements.txt
pyspark
kafka-python
yfinance
pandas
numpy
requests
python-dotenv
psycopg2-binary
cassandra-driver